Calculate rather than track allocs
The memtracker will track all allocations, even those from the standard
library. But only student code is "stamped" with line information.
Sometimes standard library allocations are never freed when a student
doesn't free a complex class type. These allocations would show in the
list of leaks, but would not be associated with file information.

The solution is to no longer track allocated bytes, but to iterate over
the list of allocations at the end. If there are any allocations that do
not have associated line information, then they are ignored.